Let's avoid all the complications that come with floating point numbers.

+++ src/main/java/com/odde/electricbartender/BeerGlass.java
@@ -4,9 +4,9 @@ package com.odde.electricbartender;
  * We'll use the metric system to avoid confusing people who aren't used to traditional/historic units.
  */
 public class BeerGlass {
-    private static double TOTAL_CAPACITY_IN_MILLITERS = 300.0;
+    private static int TOTAL_CAPACITY_IN_MILLITERS = 300;
 
-    public double getTotalCapacityInMilliters() {
+    public int getTotalCapacityInMilliters() {
         return TOTAL_CAPACITY_IN_MILLITERS;
     }
 }
